renamed folder for zustand stores  - old folder named state  - new folder now named store
